moved insert_int_sorted_list from geometry.arrow_intersect to geometry._utils; added method insert_into_sorted_list_with_indexes; added tests
